refactor(network): Centralize network and Retrofit configurations in `app` module
This commit moves network-related Dagger Hilt modules from the `data:retrofit` module to the `app` module. This centralizes the configuration for OkHttp and Retrofit, making it easier to manage and ensuring consistency across the application.

- Removed `ArticleApiModule` from `data:retrofit`.
- Added `NetworkModule` to the `app` module to provide a singleton `OkHttpClient` instance. This `OkHttpClient` is configured with a `HttpLoggingInterceptor` (logging level BODY) and an interceptor to add a `Cache-Control: no-cache` header to all requests.
- Added `RetrofitModule` to the `app` module. This module provides:
    - A binding for `ArticleApi` using `RetrofitArticleApi` (qualified with `@RemoteArticleSource`).
    - A singleton `Json` instance configured to ignore unknown keys.
    - A singleton `KesiAndroidService` instance, built using the provided `Json`, `OkHttpClient`, and `@KesiAndroidApiUrl`.
- Updated `ImageLoaderModule` in the `app` module to directly inject the `OkHttpClient` provided by `NetworkModule`, removing its own OkHttp client creation logic.
- Added necessary dependencies for Retrofit and Kotlinx Serialization to the `app` module's `build.gradle.kts`.
