+++ app/src/main/kotlin/com/kesicollection/kesiandroid/di/NetworkModule.kt
@@ -1,0 +1,28 @@
+package com.kesicollection.kesiandroid.di
+
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import okhttp3.OkHttpClient
+import okhttp3.logging.HttpLoggingInterceptor
+import javax.inject.Singleton
+
+
+@Module
+@InstallIn(SingletonComponent::class)
+object NetworkModule {
+
+    @Singleton
+    @Provides
+    fun providesOkHttpClient(): OkHttpClient = OkHttpClient.Builder()
+        .addInterceptor(HttpLoggingInterceptor().apply {
+            setLevel(HttpLoggingInterceptor.Level.BODY)
+        })
+        .addInterceptor { chain ->
+            val request = chain.request().newBuilder()
+                .header("Cache-Control", "no-cache")
+                .build()
+            chain.proceed(request)
+        }.build()
+}
